fix: 修改时间格式布局
修改时间格式布局

Log: 修改时间格式布局
Influence: 时间格式布局
Bug: https://pms.uniontech.com/bug-view-152835.html
